Extract create backmerge PR methods in helper files
